chore: Update project interface and contact interface
The project interface (interfaceProjects.ts) was modified to include a new property "CodeLink" to store the link to the code repository for each project. Additionally, the contact interface (interfaceContact.ts) was created to define the structure of personal contact information. These changes improve the data model and provide more flexibility for displaying project details and personal information.

Signed-off-by: lambasid <lambasid@sheridancollege.ca>